+++ packages/xgplayer-mobilevideo/src/index.js
@@ -78,10 +78,12 @@ class MobileVideo extends HTMLElement {
     this.vCtx = new VideoCtx(Object.assign({
       canvas: this._canvas
     }, { style: { width: this.width, height: this.height } }));
+    if (!this.noAudio) {
+      this.aCtx = new AudioCtx({
+        volume
+      });
+    }
 
-    this.aCtx = new AudioCtx({
-      volume
-    });
     this.ticker = new (getTicker())()
     this.reconciler = new AVReconciler({
       vCtx: this.vCtx,
@@ -96,8 +98,9 @@ class MobileVideo extends HTMLElement {
       }
       this.dispatchEvent(new Event('canplay'));
     }
-
-    this.aCtx.on('AUDIO_SOURCE_END', this.handleAudioSourceEnd)
+    if (!this.noAudio) {
+      this.aCtx.on('AUDIO_SOURCE_END', this.handleAudioSourceEnd)
+    }
   }
 
   handleAudioSourceEnd () {
@@ -112,9 +115,10 @@ class MobileVideo extends HTMLElement {
   destroy () {
     this.videoMetaInited = false;
     this.audioMetaInited = false;
-
-    this.aCtx.on('AUDIO_SOURCE_END', this.handleAudioSourceEnd)
-    this.aCtx.destroy()
+    if (!this.noAudio) {
+      this.aCtx.on('AUDIO_SOURCE_END', this.handleAudioSourceEnd)
+      this.aCtx.destroy()
+    }
     this.vCtx.destroy()
     this.ticker.stop()
     this.start = null;
@@ -125,12 +129,16 @@ class MobileVideo extends HTMLElement {
   }
 
   onDemuxComplete (videoTrack, audioTrack) {
-    // MobileVideo.resolveVideoGOP(videoTrack)
-    this.aCtx.decodeAudio(audioTrack);
+    if (!this.noAudio) {
+      this.aCtx.decodeAudio(audioTrack);
+    }
     this.vCtx.decodeVideo(videoTrack);
   }
 
   setAudioMeta (meta) {
+    if (this.noAudio) {
+      return;
+    }
     if (this.audioMetaInited) {
       this.aCtx.destroy();
       this.aCtx = new AudioCtx({});
@@ -148,10 +156,6 @@ class MobileVideo extends HTMLElement {
     }
     this.vCtx.setVideoMetaData(meta);
     this.videoMetaInited = true;
-  }
-
-  disconnectedCallback () {
-    this.destroy();
   }
 
   get width () {
@@ -230,7 +234,9 @@ class MobileVideo extends HTMLElement {
 
   set playbackRate (val) {
     this.setAttribute('playbackrate', val);
-    this.aCtx.playbackRate = val;
+    if (!this.noAudio) {
+      this.aCtx.playbackRate = val;
+    }
     this.vCtx.playbackRate = val;
 
     this.dispatchEvent(new Event('ratechange'))
@@ -262,10 +268,15 @@ class MobileVideo extends HTMLElement {
       this.destroy()
       this.init()
     }
-    this.dispatchEvent(new Event('play'))
+    let audioPlayTask = null
+    if (this.noAudio) {
+      audioPlayTask = Promise.resolve()
+    } else {
+      audioPlayTask = this.aCtx.play()
+    }
     this.pendingPlayTask = Promise.all([
       this.vCtx.play(),
-      this.aCtx.play().then(() => {
+      audioPlayTask.then(() => {
         // this.aCtx.muted = true
       })
     ]).then(() => {
@@ -291,7 +302,10 @@ class MobileVideo extends HTMLElement {
             this.dispatchEvent(new Event('waiting'))
           }
         }
-      })
+        if (this.noAudio) {
+          this.vCtx.cleanBuffer();
+        }
+      });
 
       this.pendingPlayTask = null
       this.played = true;
@@ -303,7 +317,9 @@ class MobileVideo extends HTMLElement {
 
   pause () {
     this._paused = true;
-    this.aCtx.pause()
+    if (!this.noAudio) {
+      this.aCtx.pause()
+    }
     this.vCtx.pause()
 
     this.dispatchEvent(new Event('pause'))
@@ -314,10 +330,16 @@ class MobileVideo extends HTMLElement {
   }
 
   get volume () {
+    if (this.noAudio) {
+      return 0;
+    }
     return this.aCtx.volume
   }
 
   set volume (vol) {
+    if (this.noAudio) {
+      return;
+    }
     this.setAttribute('volume', vol);
     this.aCtx.volume = vol
     if (vol > 0 && this.muted) {
@@ -338,6 +360,9 @@ class MobileVideo extends HTMLElement {
   }
 
   set muted (val) {
+    if (this.noAudio) {
+      return;
+    }
     this.setAttribute('muted', val);
     if (!val) {
       this.aCtx.muted = false
@@ -348,11 +373,15 @@ class MobileVideo extends HTMLElement {
   }
 
   get error () {
-    return this.vCtx.error || this.aCtx.error;
+    return this.vCtx.error || (this.noAudio ? null : this.aCtx.error);
   }
 
   get buffered () {
     return this.vCtx.buffered
+  }
+
+  get noAudio () {
+    return this.getAttribute('noaudio') === 'true'
   }
 }
 // eslint-disable-next-line no-undef
